+++ src/io/forsta/ccsm/util/NetworkUtils.java
@@ -1,4 +1,4 @@
-package io.forsta.util;
+package io.forsta.ccsm.util;
 
 import android.util.Log;
 import org.json.JSONException;
